# frontend/src/api/generated/personnel.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DutyFlow personnel",
    "version": "0.1.0"
  },
  "paths": {
...
    "/internal/people/batch": {
      "post": {
        "tags": [
          "internal"
        ],
        "summary": "People Batch",
        "operationId": "people_batch_internal_people_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PeopleBatchIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PeopleBatchPerson"
                  },
                  "type": "array",
                  "title": "Response People Batch Internal People Batch Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/people/availability-batch": {
      "post": {
        "tags": [
          "internal"
        ],
        "summary": "Availability Batch",
        "description": "Доступность по дням: строка '1'/'0' на каждого человека. Архивные — все '0'.",
        "operationId": "availability_batch_internal_people_availability_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvailabilityIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AvailabilityIn": {
        "properties": {
          "person_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 50000,
            "minItems": 1,
            "title": "Person Ids"
          },
          "date_from": {
            "type": "string",
            "format": "date",
            "title": "Date From"
          },
          "date_to": {
            "type": "string",
            "format": "date",
            "title": "Date To"
          }
        },
        "type": "object",
        "required": [
          "person_ids",
          "date_from",
          "date_to"
        ],
        "title": "AvailabilityIn"
      },
      "AvailabilityOut": {
        "properties": {
          "date_from": {
            "type": "string",
            "format": "date",
            "title": "Date From"
          },
          "days": {
            "type": "integer",
            "title": "Days"
          },
          "masks": {
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "format": "uuid"
            },
            "type": "object",
            "title": "Masks"
          }
        },
        "type": "object",
        "required": [
          "date_from",
          "days",
          "masks"
        ],
        "title": "AvailabilityOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PeopleBatchIn": {
        "properties": {
          "unit_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 10000,
            "title": "Unit Ids"
          },
          "include_descendants": {
            "type": "boolean",
            "title": "Include Descendants",
            "default": true
          },
          "person_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 50000,
            "title": "Person Ids"
          },
          "date_from": {
            "type": "string",
            "format": "date",
            "title": "Date From"
          },
          "date_to": {
            "type": "string",
            "format": "date",
            "title": "Date To"
          },
          "include_names": {
            "type": "boolean",
            "title": "Include Names",
            "default": false
          },
          "duty_role_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 1000,
            "title": "Duty Role Ids"
          }
        },
        "type": "object",
        "required": [
          "date_from",
          "date_to"
        ],
        "title": "PeopleBatchIn",
        "description": "Выборка — либо по подразделениям (действующие люди), либо по конкретным людям\n(включая исключённых из списков: scheduling проверяет свои назначения)."
      },
      "PeopleBatchPerson": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "unit_id": {
            "type": "string",
            "format": "uuid",
            "title": "Unit Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "rank_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank Id"
          },
          "rank_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank Order"
          },
          "position_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position Id"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "attributes": {
            "additionalProperties": true,
            "type": "object",
            "title": "Attributes"
          },
          "exemptions": {
            "items": {
              "prefixItems": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "string",
                  "format": "date"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Exemptions"
          },
          "clearances": {
            "items": {
              "prefixItems": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "anyOf": [
                    {
                      "type": "string",
                      "format": "date"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                {
                  "anyOf": [
                    {
                      "type": "string",
                      "format": "date"
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              ],
              "type": "array",
              "maxItems": 3,
              "minItems": 3
            },
            "type": "array",
            "title": "Clearances"
          },
          "last_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Name"
          },
          "first_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Name"
          },
          "middle_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Middle Name"
          },
          "rank_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "unit_id",
          "is_active",
          "rank_id",
          "rank_order",
          "position_id",
          "attributes",
          "exemptions",
          "clearances"
        ],
        "title": "PeopleBatchPerson"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}